% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{BusinessReport-package}
\alias{BusinessReport}
\alias{BusinessReport-package}
\title{BusinessReport: Modelos profissionais de relatórios para Quarto/Typst}
\description{
Fornece modelos profissionais de relatórios em Quarto com renderização via
Typst para usos corporativos, analíticos e técnicos. Inclui capa e
contracapa opcionais, uma coleção curada de fontes e três estilos distintos
de sumário.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lemuelemos.github.io/BusinessReports/}
  \item \url{https://github.com/lemuelemos/BusinessReports}
  \item Report bugs at \url{https://github.com/lemuelemos/BusinessReports/issues}
}

}
\author{
\strong{Maintainer}: Lemuel Lemos \email{lemuelemos@gmail.com}

}
\keyword{internal}
