% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{open_report}
\alias{open_report}
\title{Abra o arquivo principal de um projeto \code{BusinessReport} na IDE}
\usage{
open_report(path = ".")
}
\arguments{
\item{path}{Diretorio raiz do projeto. O padrao e \code{"."}.}
}
\value{
\code{path}, invisivelmente.
}
\description{
Abra o arquivo principal de um projeto \code{BusinessReport} na IDE
}
