% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toc.R
\name{toggle_back_cover}
\alias{toggle_back_cover}
\title{Ative ou desative a contracapa de um projeto \code{BusinessReport} existente}
\usage{
toggle_back_cover(path = ".", back_cover)
}
\arguments{
\item{path}{Diretório raiz do projeto.}

\item{back_cover}{Lógico. \code{TRUE} para incluir a contracapa; \code{FALSE} para
removê-la.}
}
\value{
\code{path}, invisivelmente.
}
\description{
Ative ou desative a contracapa de um projeto \code{BusinessReport} existente
}
\examples{
\dontrun{
toggle_back_cover("meu-relatorio", back_cover = FALSE)
}

}
