% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toc.R
\name{toggle_cover}
\alias{toggle_cover}
\title{Ative ou desative a capa de um projeto \code{BusinessReport} existente}
\usage{
toggle_cover(path = ".", cover)
}
\arguments{
\item{path}{Diretório raiz do projeto.}

\item{cover}{Lógico. \code{TRUE} para incluir a capa; \code{FALSE} para removê-la.}
}
\value{
\code{path}, invisivelmente.
}
\description{
Ative ou desative a capa de um projeto \code{BusinessReport} existente
}
\examples{
\dontrun{
toggle_cover("meu-relatorio", cover = FALSE)
}

}
