% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{BusinessReport-package}
\alias{BusinessReport}
\alias{BusinessReport-package}
\title{BusinessReport: Modelos profissionais de relatorios para Quarto/Typst}
\description{
Fornece modelos profissionais de relatorios em Quarto com renderizacao via
Typst para usos corporativos, analiticos e tecnicos. Inclui capa e
contracapa opcionais, uma colecao curada de fontes e quatro estilos
de sumario.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lemuelemos.github.io/BusinessReports/}
  \item \url{https://github.com/lemuelemos/BusinessReports}
  \item Report bugs at \url{https://github.com/lemuelemos/BusinessReports/issues}
}

}
\author{
\strong{Maintainer}: Lemuel Lemos \email{lemuelemos@gmail.com}

}
\keyword{internal}
