% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toc.R
\name{list_toc_styles}
\alias{list_toc_styles}
\title{Liste os estilos de sumario disponiveis em \code{BusinessReport}}
\usage{
list_toc_styles()
}
\value{
Um data frame com as colunas \code{id}, \code{name} e \code{description}.
}
\description{
Retorna um data frame descrevendo cada estilo de sumario. Passe o valor de
\code{id} (1, 2, 3 ou 4) para \code{\link[=create_business_report]{create_business_report()}} ou \code{\link[=set_toc_style]{set_toc_style()}}.
}
\examples{
list_toc_styles()

}
